Replaced short open tags with long form open tags (i.e. "<?php").
This was done to make the code compatible on all servers, including
those with "short_open_tag" turned off in php.ini.  The short echo
syntax (i.e. "<?=") also does not work when short codes are disabled in
versions of PHP prior to 5.4.0.  This was the default environment on a
popular web host I use, so I was unable to use this script without these
changes, which shouldn't have any negative impact on any environment.  I
also made changes to the spacing before a few end tags to make things
consistent throughout the script.